Change custom callback logic to layers.c
