+++ keyboards/gmmk/pro/rev1/ansi/keymaps/michael-mml/def.h
@@ -14,6 +14,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
+#pragma once
 
 enum LAYER_NAME_TO_LAYER_INDEX_MAP {
     WIN,    // defaults to 0 and increments by 1
