+++ keyboards/gmmk/pro/rev1/ansi/keymaps/michael-mml/keymap.c
@@ -18,19 +18,12 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "keycodes.h"
 #include QMK_KEYBOARD_H
 #include "def.h"
+#include "layers.c"
 
 /**
  * This file was generated by qmk json2c and edited to support the rotary
  * encoder and rgb.
  */
-
-static void disableSideLeds(uint8_t led_min, uint8_t led_max, int *leftSideLedIndices, int *rightSideLedIndices) {
-    for (int ledIndex = 0; ledIndex < 8; ledIndex++) {
-        RGB_MATRIX_INDICATOR_SET_COLOR(leftSideLedIndices[ledIndex], 0, 0, 0);
-        RGB_MATRIX_INDICATOR_SET_COLOR(rightSideLedIndices[ledIndex], 0, 0, 0);
-    }
-}
-
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
     // clang-format off
     //      ESC      F1       F2       F3       F4       F5       F6       F7       F8       F9       F10      F11      F12	     Prt           Rotary(Mute)
@@ -127,47 +120,11 @@ const uint16_t PROGMEM encoder_map[][NUM_ENCODERS][NUM_DIRECTIONS] = {
 };
 #endif
 
-// TODO: this function is probably outdated
-void determine_rgb_per_layer(uint8_t led_min, uint8_t led_max) {
-    int  LEFT_SIDE_LED_INDICES[8]  = {67, 70, 73, 76, 80, 83, 87, 91};
-    int  RIGHT_SIDE_LED_INDICES[8] = {68, 71, 74, 77, 81, 84, 88, 92};
-    bool isCapsLockEnabled         = host_keyboard_led_state().caps_lock;
-    bool isNumLockEnabled          = host_keyboard_led_state().num_lock;
-    switch (get_highest_layer(layer_state)) {
-        case WIN: {
-            if (isCapsLockEnabled) {
-                RGB_MATRIX_INDICATOR_SET_COLOR(3, 255, 0, 0);
-            }
-            if (isNumLockEnabled) {
-                RGB_MATRIX_INDICATOR_SET_COLOR(40, 30, 252, 131);
-            }
-            disableSideLeds(led_min, led_max, LEFT_SIDE_LED_INDICES, RIGHT_SIDE_LED_INDICES);
-            break;
-        }
-        case WIN_FN: {
-            rgb_matrix_set_color_all(30, 252, 131);
-            break;
-        }
-        case MAC: {
-            if (isCapsLockEnabled) {
-                RGB_MATRIX_INDICATOR_SET_COLOR(3, 0, 30, 255);
-            }
-            if (isNumLockEnabled) {
-                RGB_MATRIX_INDICATOR_SET_COLOR(40, 207, 207, 207);
-            }
-            disableSideLeds(led_min, led_max, LEFT_SIDE_LED_INDICES, RIGHT_SIDE_LED_INDICES);
-            break;
-        }
-        case MAC_FN: {
-            rgb_matrix_set_color_all(207, 207, 207);
-            break;
-        }
-        default:
-            break;
-    }
-}
-
 bool rgb_matrix_indicators_advanced_user(uint8_t led_min, uint8_t led_max) {
     determine_rgb_per_layer(led_min, led_max);
+    return true;
+}
+
+bool encoder_update_user(uint8_t index, bool clockwise) {
     return true;
 }
